+++ src/main/java/com/conspring/banco/application/services/UserService.java
@@ -3,6 +3,7 @@ package com.conspring.banco.application.services;
 import com.conspring.banco.api.dtos.AccountDto;
 import com.conspring.banco.api.mappers.AccountMapper;
 import com.conspring.banco.api.mappers.UserMapper;
+import com.conspring.banco.domain.exceptions.NoSeEncontroE;
 import com.conspring.banco.domain.models.Account;
 import com.conspring.banco.domain.models.User;
 import com.conspring.banco.api.dtos.UserDto;
@@ -11,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -41,18 +43,41 @@ public class UserService {
         return userDto;
     }
 
-    public UserDto modificarUser(UserDto userDto) {
-        User user = userRepository.save(User.builder()
-                .id(userDto.getId())
-                .username(userDto.getUsername())
-                .password(userDto.getPassword())
-                .build());
-        userDto = UserMapper.UserToDtoMap(user);
-        return userDto;
+
+    public User buscarPorId(Integer id) throws NoSeEncontroE {
+        if (id != null){
+            User user = userRepository.findById(id).orElse(null);
+            return user;
+        } else {
+            throw new NoSeEncontroE("Debe ingresar un ID");
+        }
+
+
     }
 
-    public void borrarById (Integer id){
-        userRepository.deleteById(id);
+    public UserDto modificarUser(UserDto userDto) throws NoSeEncontroE {
+
+            if (buscarPorId(userDto.getId()) != null){
+                User user = userRepository.save(User.builder()
+                        .id(userDto.getId())
+                        .username(userDto.getUsername())
+                        .password(userDto.getPassword())
+                        .build());
+                userDto = UserMapper.UserToDtoMap(user);
+                return userDto;
+            } else {
+                throw new  NoSeEncontroE("No existe el usuario con ID " + userDto.getId());
+            }
+    }
+
+    public String borrarById (Integer id){
+        if (userRepository.existsById(id)){
+            userRepository.deleteById(id);
+            return "Usuario eliminado exitosamente";
+        } else {
+            return "El ID a eliminar no existe";
+        }
+
     }
